# swagger.json
{
	"components": {
...
		"schemas": {
...
			"SellerType": {
				"enum": [
					"ARCHITECT",
					"COMPANY"
				],
				"type": "string"
			},
			"ISellerRegister": {
				"properties": {
					"sellerType": {
						"$ref": "#/components/schemas/SellerType"
					},
					"identityCardNumber": {
						"type": "string"
					},
					"identityCardDate": {
						"type": "string",
						"format": "date-time"
					},
					"identityCardPlace": {
						"type": "string"
					},
					"taxCode": {
						"type": "string"
					},
					"bankAccountNumber": {
						"type": "string"
					},
					"bankAccountName": {
						"type": "string"
					},
					"bankName": {
						"type": "string"
					},
					"bankBranch": {
						"type": "string"
					}
				},
				"required": [
					"sellerType",
					"identityCardNumber",
					"identityCardDate",
					"identityCardPlace",
					"taxCode",
					"bankAccountNumber",
					"bankAccountName",
					"bankName",
					"bankBranch"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ISellerUpdate": {
				"properties": {
					"bankAccountNumber": {
						"type": "string"
					},
					"bankAccountName": {
						"type": "string"
					},
					"bankName": {
						"type": "string"
					},
					"bankBranch": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"IShopStatusUpdate": {
				"properties": {
					"shopStatus": {
						"type": "boolean"
					}
				},
				"required": [
					"shopStatus"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "vro_backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/sellers/register": {
			"post": {
				"operationId": "CreateSeller",
				"responses": {
					"200": {
						"description": "200 - Return message and status",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"sellerType": "ARCHITECT",
											"identityCardNumber": "123456789",
											"identityCardDate": "2021-01-01T00:00:00.000Z",
											"identityCardPlace": "Hà Nội",
											"taxCode": "123456789",
											"bankAccountNumber": "123456789",
											"bankAccountName": "Nguyễn Văn A",
											"bankName": "Vietcombank",
											"bankBranch": "Hà Nội"
										}
									}
								}
							}
						}
					}
				},
				"description": "sellerType: ARCHITECT, COMPANY",
				"summary": "REGISTER NEW SELLER REQUEST (Only user can register seller)",
				"tags": [
					"Sellers"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "- Seller's info",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISellerRegister",
								"description": "- Seller's info"
							}
						}
					}
				}
			}
		},
		"/sellers/registration-form": {
			"get": {
				"operationId": "GetRegistrationForm",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"summary": "GET ALL REGISTRATION FORM TO CONFIRM (Only admin can get registration form of all users)",
				"tags": [
					"Sellers"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- Number of items per page",
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- Number of skipped items",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "isApproved",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/sellers/approve/{id}": {
			"put": {
				"operationId": "ApproveSeller",
				"responses": {
					"200": {
						"description": "200 - Return message and status",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"summary": "APPROVE SELLER (Only admin can approve seller)",
				"tags": [
					"Sellers"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- Seller's id",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/sellers/profile": {
			"get": {
				"operationId": "GetSellerProfile",
				"responses": {
					"200": {
						"description": "200 - Return seller's profile",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"summary": "GET SELLER PROFILE (Only seller can get their profile)",
				"tags": [
					"Sellers"
				],
				"security": [
					{
						"jwt": [
							"seller"
						]
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateSellerProfile",
				"responses": {
					"200": {
						"description": "200 - Return message and status",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"summary": "UPDATE SELLER PROFILE (Only seller can update their profile)",
				"tags": [
					"Sellers"
				],
				"security": [
					{
						"jwt": [
							"seller"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "- Seller's info",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISellerUpdate",
								"description": "- Seller's info"
							}
						}
					}
				}
			}
		},
		"/sellers/shop-status": {
			"put": {
				"operationId": "UpdateShopStatus",
				"responses": {
					"200": {
						"description": "200 - Return message and status",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"summary": "OPEN/CLOSE SHOP (Only seller can)",
				"tags": [
					"Sellers"
				],
				"security": [
					{
						"jwt": [
							"seller"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "- Shop's status",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IShopStatusUpdate",
								"description": "- Shop's status"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}